requirements.txt
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.4.0
pymysql==1.1.0
requests==2.31.0
